Refactor panel, fix opening
