+++ packages/grid/_modules_/grid/components/tools/index.ts
@@ -1,5 +1,7 @@
+export * from './ColumnsPanel';
 export * from './FilterForm';
-export * from './FilterPanel';
-export * from './Preferences';
-export * from './FilterInputValueProps';
 export * from './FilterInputValue';
+export * from './FilterInputValueProps';
+export * from './FilterPanel';
+export * from './Panel';
+export * from './Preferences';
